[Fix] (Breaking Change) better typing

+++ src/moneywiz_api/managers/investment_holding_manager.py
@@ -1,4 +1,5 @@
 from typing import Dict, Callable, List
+from decimal import Decimal
 
 from moneywiz_api.model.investment_holding import InvestmentHolding
 from moneywiz_api.managers.record_manager import RecordManager
@@ -18,8 +19,8 @@ class InvestmentHoldingManager(RecordManager[InvestmentHolding]):
     def get_holdings_for_account(self, account_id: ID) -> List[InvestmentHolding]:
         return [x for _, x in self.records().items() if x.account == account_id]
 
-    def update_last_price(self, latest_price: float):
+    def update_last_price(self, latest_price: Decimal):
         raise NotImplementedError()
 
-    def update_price_table(self, latest_price: float):
+    def update_price_table(self, latest_price: Decimal):
         raise NotImplementedError()
